+++ drift_dev/lib/src/analysis/resolver/drift/element_resolver.dart
@@ -4,7 +4,6 @@ import 'package:collection/collection.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:sqlparser/sqlparser.dart';
-import 'package:sqlparser/utils/find_referenced_tables.dart';
 
 import '../../backend.dart';
 import '../../driver/error.dart';
@@ -174,70 +173,8 @@ abstract class DriftElementResolver<T extends DiscoveredElement>
     }
   }
 
-  Future<SqlEngine> newEngineWithTables(
-      Iterable<DriftElement> references) async {
-    final mapping = await resolver.driver.typeMapping;
-    return mapping.newEngineWithTables(references);
-  }
-
   DriftElement? findInResolved(List<DriftElement> references, String name) {
     return references.firstWhereOrNull((e) => e.id.sameName(name));
-  }
-
-  Future<List<DriftElement>> resolveTableReferences(AstNode stmt) async {
-    final engine = resolver.driver.newSqlEngine();
-    final references = engine.findReferencedSchemaTables(stmt);
-    final found = <DriftElement>[];
-    final missingNames = <String, ResolveReferencedElementResult>{};
-
-    for (final table in references) {
-      // If this is a reference to a table the empty engine already knows, it
-      // must be a table builtin to sqlite3, not a drift reference.
-      if (engine.knownResultSets
-          .any((e) => e.name.toLowerCase() == table.toLowerCase())) {
-        continue;
-      }
-
-      final result = await resolver.resolveReference(discovered.ownId, table);
-
-      if (result is ResolvedReferenceFound) {
-        found.add(result.element);
-      } else {
-        missingNames[table.toLowerCase()] = result;
-      }
-    }
-
-    if (missingNames.isNotEmpty) {
-      // Ok, there are unresolved table references
-      for (final reference in stmt.allDescendants.whereType<TableReference>()) {
-        if (reference.resolved == null) {
-          final unresolvedBecause =
-              missingNames[reference.tableName.toLowerCase()];
-
-          if (unresolvedBecause != null) {
-            reportErrorForUnresolvedReference(unresolvedBecause,
-                (msg) => DriftAnalysisError.inDriftFile(reference, msg));
-          }
-        }
-      }
-    }
-
-    return found;
-  }
-
-  /// Finds all referenced tables, Dart expressions and Dart types referenced
-  /// in [stmt].
-  Future<FoundReferencesInSql> resolveSqlReferences(AstNode stmt) async {
-    final driftElements = await resolveTableReferences(stmt);
-
-    final identifier = _IdentifyDartElements();
-    stmt.accept(identifier, null);
-
-    return FoundReferencesInSql(
-      referencedElements: driftElements,
-      dartExpressions: identifier.dartExpressions,
-      dartTypes: identifier.dartTypes,
-    );
   }
 
   /// Creates a type resolver capable of resolving `ENUM` and `ENUMNAME` types.
@@ -285,31 +222,4 @@ class FoundReferencesInSql {
 
   static final RegExp enumRegex =
       RegExp(r'^enum(name)?\((\w+)\)$', caseSensitive: false);
-}
-
-class _IdentifyDartElements extends RecursiveVisitor<void, void> {
-  final List<String> dartExpressions = [];
-  final Map<SyntacticEntity, String> dartTypes = {};
-
-  @override
-  void visitCastExpression(CastExpression e, void arg) {
-    final match = FoundReferencesInSql.enumRegex.firstMatch(e.typeName);
-
-    if (match != null) {
-      // Found `ENUMNAME(x)`, where `x` is a Dart type that we might want to
-      // resolve later.
-      dartTypes[e] = match.group(2)!;
-    }
-
-    super.visitCastExpression(e, arg);
-  }
-
-  @override
-  void visitColumnConstraint(ColumnConstraint e, void arg) {
-    if (e is MappedBy) {
-      dartExpressions.add(e.mapper.dartCode);
-    } else {
-      super.visitColumnConstraint(e, arg);
-    }
-  }
 }